Trivial refactoring: remove unnecessary boxing and unboxing for simplicity(closes #862)
(cherry picked from commit 12f0a099d046b9ed49464626d823237d5efca3fc)

+++ src/main/java/org/codehaus/groovy/runtime/DefaultGroovyMethods.java
@@ -16096,7 +16096,7 @@ public class DefaultGroovyMethods extends DefaultGroovyMethodsSupport {
      * @since 1.0
      */
     public static long abs(Long number) {
-        return Math.abs(number.longValue());
+        return Math.abs(number);
     }
 
     /**
@@ -16107,7 +16107,7 @@ public class DefaultGroovyMethods extends DefaultGroovyMethodsSupport {
      * @since 1.0
      */
     public static float abs(Float number) {
-        return Math.abs(number.floatValue());
+        return Math.abs(number);
     }
 
     /**
@@ -16129,7 +16129,7 @@ public class DefaultGroovyMethods extends DefaultGroovyMethodsSupport {
      * @since 1.0
      */
     public static int round(Float number) {
-        return Math.round(number.floatValue());
+        return Math.round(number);
     }
 
     /**
